Draft - working through some kind of cleanup error.
