+++ core/node/rpc/archiver_test.go
@@ -244,14 +244,14 @@ func TestArchiveOneStream(t *testing.T) {
 	)
 	require.NoError(err)
 
-	dbCfg, schema, schemaDeleter, err := dbtestutils.StartDB(ctx)
+	dbCfg, schema, schemaDeleter, err := dbtestutils.ConfigureDB(ctx)
 	require.NoError(err)
 	defer schemaDeleter()
 
 	pool, err := storage.CreateAndValidatePgxPool(ctx, dbCfg, schema, nil)
 	require.NoError(err)
 
-	streamStorage, err := storage.NewPostgresEventStore(
+	streamStorage, err := storage.NewPostgresStreamStore(
 		ctx,
 		pool,
 		GenShortNanoid(),
