+++ core/node/rpc/render/templates/debug/multi.template.html
@@ -138,7 +138,6 @@ UTC: {{.StartTime}}</pre>
           <span class="tooltiptext">
             Version: {{.PostgresStatus.Version}}
             SystemId: {{.PostgresStatus.SystemId}}
-            EsCount: {{.PostgresStatus.EsCount}}
             Total Connections: {{.PostgresStatus.TotalConns}}
             Acquired Connections: {{.PostgresStatus.AcquiredConns}}
             Idle Connections: {{.PostgresStatus.IdleConns}}
