+++ core/node/storage/storage.go
@@ -2,6 +2,7 @@ package storage
 
 import (
 	"context"
+
 	"github.com/ethereum/go-ethereum/common"
 
 	. "github.com/river-build/river/core/node/shared"
